+++ exportplan/data.py
@@ -30,6 +30,17 @@ if settings.FEATURE_EXPORT_PLAN_SECTIONS_DISABLED:
 else:
     SECTIONS_DISABLED = []
 
+# This maintains a collection of which pages require a country to be selected
+COUNTRY_REQUIRED = [
+    slugify(TARGET_MARKETS_RESEARCH),
+    slugify(ADAPTATION_TARGET_MARKET),
+    slugify(MARKETING_APPROACH),
+    slugify(COSTS_AND_PRICING),
+    slugify(FINANCE),
+    slugify(PAYMENT_METHODS),
+    slugify(TRAVEL_AND_BUSINESS_POLICIES),
+]
+
 SECTIONS = OrderedDict({
     slugify(ABOUT_YOUR_BUSINESS): {
         'title': ABOUT_YOUR_BUSINESS,