+++ CHANGELOG.md
@@ -258,3 +258,5 @@
 - noticket bug fix lessons details on objectives
 - GP2-706 - 
 - ablity-to-disable-export-sections
+- GP2-709 - BE Country Model
+